Updating createApp test command
In some cases Cypress is running too fast, meaning that the test runs ahead of what is actual seen on Budibase.

This issue mainly occurs at the start of the smoke build and around logging in.

I am attempting to combat this by ensuring the url includes 'apps' after logging in